+++ src/index.d.ts
@@ -214,6 +214,15 @@ import {
 } from './services/liveTesting.ts';
 
 import {
+  acceptInvite,
+  createAccount,
+  createInvites,
+  fetchUsersMultiAccountDetails,
+  removeUserFromActiveMultiUserAccount,
+  rescindInvite
+} from './services/multi-user-accounts/multi-user-accounts.ts';
+
+import {
 	calculateLongestStreaksOffline,
 	getPracticeSessionsOffline,
 	getRecentActivityOffline,
@@ -323,6 +332,7 @@ import {
 	getSanityDate,
 	getSongTypesFor,
 	getSortOrder,
+	hasAnyMethodV2IntroCompleted,
 	jumpToContinueContent
 } from './services/sanity.js';
 
@@ -457,11 +467,12 @@ import {
 } from './services/userActivity.js';
 
 import {
-	 default as EventsAPI 
+	 default as EventsAPI
 } from './services/eventsAPI';
 
 declare module 'musora-content-services' {
 	export {
+		acceptInvite,
 		addContextToContent,
 		addContextToLearningPaths,
 		addItemToPlaylist,
@@ -485,8 +496,10 @@ declare module 'musora-content-services' {
 		contentStatusReset,
 		contentStatusStarted,
 		convertToTimeZone,
+		createAccount,
 		createComment,
 		createForumCategory,
+		createInvites,
 		createPlaylist,
 		createPost,
 		createPracticeNotes,
@@ -602,6 +615,7 @@ declare module 'musora-content-services' {
 		fetchUserPracticeMeta,
 		fetchUserPracticeNotes,
 		fetchUserPractices,
+		fetchUsersMultiAccountDetails,
 		filterCoursesInCourseCollections,
 		flushWatchSession,
 		followThread,
@@ -678,6 +692,7 @@ declare module 'musora-content-services' {
 		getWeekNumber,
 		globalConfig,
 		guidedCourses,
+		hasAnyMethodV2IntroCompleted,
 		initializeEnvVar,
 		initializeService,
 		isBucketUrl,
@@ -724,12 +739,14 @@ declare module 'musora-content-services' {
 		registerProgressCallback,
 		removeContentAsInterested,
 		removeContentAsNotInterested,
+		removeUserFromActiveMultiUserAccount,
 		removeUserPractice,
 		replyToComment,
 		report,
 		reportComment,
 		reportPlaylist,
 		requestEmailChange,
+		rescindInvite,
 		reset,
 		resetAllAwards,
 		resetAllLearningPaths,