+++ src/index.js
@@ -1,7 +1,7 @@
 /*** This file was generated automatically. To recreate, please run `npm run build-index`. ***/
 
 import {
-	 default as EventsAPI 
+	 default as EventsAPI
 } from './services/eventsAPI';
 
 import {
@@ -227,6 +227,15 @@ import {
 } from './services/offline/offline.ts';
 
 import {
+	acceptInvite,
+	createAccount,
+	createInvites,
+	fetchUsersMultiAccountDetails,
+	removeUserFromActiveMultiUserAccount,
+	rescindInvite
+} from './services/multi-user-accounts/multi-user-accounts.ts';
+
+import {
 	emitProgressSaved,
 	onProgressSaved
 } from './services/progress-events.js';
@@ -327,6 +336,7 @@ import {
 	getSanityDate,
 	getSongTypesFor,
 	getSortOrder,
+	hasAnyMethodV2IntroCompleted,
 	jumpToContinueContent
 } from './services/sanity.js';
 
@@ -461,6 +471,7 @@ import {
 } from './services/userActivity.js';
 
 export {
+	acceptInvite,
 	addContextToContent,
 	addContextToLearningPaths,
 	addItemToPlaylist,
@@ -484,8 +495,10 @@ export {
 	contentStatusReset,
 	contentStatusStarted,
 	convertToTimeZone,
+	createAccount,
 	createComment,
 	createForumCategory,
+	createInvites,
 	createPlaylist,
 	createPost,
 	createPracticeNotes,
@@ -601,6 +614,7 @@ export {
 	fetchUserPracticeMeta,
 	fetchUserPracticeNotes,
 	fetchUserPractices,
+	fetchUsersMultiAccountDetails,
 	filterCoursesInCourseCollections,
 	flushWatchSession,
 	followThread,
@@ -677,6 +691,7 @@ export {
 	getWeekNumber,
 	globalConfig,
 	guidedCourses,
+	hasAnyMethodV2IntroCompleted,
 	initializeEnvVar,
 	initializeService,
 	isBucketUrl,
@@ -723,12 +738,14 @@ export {
 	registerProgressCallback,
 	removeContentAsInterested,
 	removeContentAsNotInterested,
+	removeUserFromActiveMultiUserAccount,
 	removeUserPractice,
 	replyToComment,
 	report,
 	reportComment,
 	reportPlaylist,
 	requestEmailChange,
+	rescindInvite,
 	reset,
 	resetAllAwards,
 	resetAllLearningPaths,